+++ src/main/kotlin/com/example/team/haribo/goms/domain/studentcouncil/service/impl/StudentCouncilMemberFilterServiceImpl.kt
@@ -16,7 +16,7 @@ class StudentCouncilMemberFilterServiceImpl(
 
     override fun filter(
         name: String?,
-        grade: Long?,
+        grade: Int?,
         department: Department?,
         status: Status?,
         role: Role?
